pyproject.toml
[tool.poetry]
name = "trading-bot-creator"
version = "0.1.0"
description = ""
authors = ["lazy-geeek <lazy-geek@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytube = "^15.0.0"
python-decouple = "^3.8"
youtube-transcript-api = "^0.6.2"
tqdm = "^4.66.2"
openai = "^1.14.3"
groq = "^0.4.2"
werkzeug = "^3.0.2"
google-api-python-client = "^2.125.0"
google-auth-oauthlib = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
